// src/swagger-telaradb/model/itemList.ts
/**
 * Telara DB
 * TelaraDB-Api
 *
 * OpenAPI spec version: 0.0.1
 * Contact: developer@tobiaskloss.com
 *
 * NOTE: This class is auto generated by the swagger-telaradb code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Item } from './item';


export interface ItemList {
    items?: Array<Item>;
    page?: number;
    pages?: number;
    total?: number;
}